Take out main app code into separate component
