+++ src/client/app/components/index.ts
@@ -5,5 +5,3 @@ export * from './command';
 export * from './snippet.info';
 export * from './profile';
 export * from './about';
-export * from './view';
-export * from './view.error';
