+++ src/client/app/containers/editor.ts
@@ -1,4 +1,4 @@
-import { Component, HostListener, AfterViewInit, ViewChild, ElementRef } from '@angular/core';
+import { Component, Input, HostListener, AfterViewInit, ViewChild, ElementRef } from '@angular/core';
 import { Dictionary } from '@microsoft/office-js-helpers';
 import * as fromRoot from '../reducers';
 import { Store } from '@ngrx/store';
@@ -21,9 +21,9 @@ import { debounce } from 'lodash';
     `
 })
 export class Editor implements AfterViewInit {
-    private _monacoEditor: monaco.editor.IStandaloneCodeEditor;
     @ViewChild('editor') private _editor: ElementRef;
-    private _readonly: boolean;
+    @Input() isViewMode: boolean;
+    private _monacoEditor: monaco.editor.IStandaloneCodeEditor;
 
     tabs = new Dictionary<IMonacoEditorState>();
     currentState: IMonacoEditorState;
@@ -40,7 +40,7 @@ export class Editor implements AfterViewInit {
      */
     async ngAfterViewInit() {
         let _overrides = { theme: 'vs' };
-        if (/#\/view/.test(location.hash)) {
+        if (this.isViewMode) {
             _overrides['readOnly'] = true;
         }
         this._monacoEditor = await this._monaco.create(this._editor, _overrides);
@@ -179,7 +179,7 @@ export class Editor implements AfterViewInit {
      * The same update happens even on tab switch.
      */
     private _debouncedInput = debounce(() => {
-        if (!this._readonly) {
+        if (!this.isViewMode) {
             this.currentState.content = this._monacoEditor.getValue();
             this._store.dispatch(new Snippet.SaveAction(this.snippet));
         }
