+++ src/client/app/containers/index.ts
@@ -1,4 +1,7 @@
 export * from './editor';
+export * from './editor.mode';
+export * from './view.mode';
+export * from './view.mode.error';
 export * from './app';
 export * from './gallery.list';
 export * from './import';
