+++ src/client/assets/styles/editor.scss
@@ -5,7 +5,7 @@
 @import 'containers';
 
 .app {
-    view {
+    view-mode {
         width: 100%;
     }
