+++ src/client/components.ts
@@ -1,10 +1,13 @@
-import { Collapse, Command, Alert, About, Dialog, SnippetInfo, Profile, ViewMode, ViewModeError  } from './app/components';
-import { Editor, GalleryList, Import } from './app/containers';
+import { Collapse, Command, Alert, About, Dialog, SnippetInfo, Profile  } from './app/components';
+import { Editor, EditorMode, GalleryList, Import, ViewMode, ViewModeError } from './app/containers';
 
 export const COMPONENT_DECLARATIONS = [
     GalleryList,
     Editor,
+    EditorMode,
     Import,
+    ViewMode,
+    ViewModeError,
     Collapse,
     Alert,
     Profile,
@@ -12,6 +15,4 @@ export const COMPONENT_DECLARATIONS = [
     SnippetInfo,
     About,
     Command,
-    ViewMode,
-    ViewModeError
 ];
